Update Main.py
After computing RSI we're left with Bollinger Bands, ATR, MACD, and Dollar Volume.

First we compute all 3 Bollinger bands separately, then we define a custom function to calculate ATR.

To calculate MACD we need to define another function, where we need to normalize the data because we're going to use it in a machine learning model and we're going to cluster the data.

After that when calculating the dollar volume it's better to divide it by 1 Million, to make it easier to comprehend.