Update Main.py
The next step is to aggregate to monthly level and filter top 150 most liquid stocks for each month.

This is done to reduce training time and experiment with the features and strategies.

We will aggregate all the technical indicators and adjusted close price, and take the end value for each month.
From the 500 stocks in S&P 500, the top 150 are selected based on 5 year rolling average Dollar Volume and the Dollar Volume and Dollar Volume Rank columns are deleted from the dataframe.