Update Main.py
The next step is to calculate monthly returns for different time horizons and add them to the feature set, because we may want to capture Time-series dynamics that reflect momentum patterns.
We also need an outlier cut-off, so for all the values above outlier threshold, they will be assigned the percentile of that threshold.

Then we  calculate returns for the following Lags 1-month, 2-months, 3-months, 6-months, 9-months, and 12-months.