Update Main.py
During this step we'll download the Fama-French Factors and Calculate Rolling Factor Betas for each stock, we want to introduce Fama-French data to estimate the exposure of our assets to commonly known risk factors, and we're going to do that using Regression and OLS model.

The 5 Fama-French factors are -
1. Market Risk
2. Size
3. Value
4. Operating Profitability
5. Investment

These factors are used to assess risk and return of portfolios. We can access the historical factor returns using the pandas-datareader library and estimate historical exposures using the Rolling OLS Regression.

After cleaning & sorting the data we can regress the End of Month Return with Factor Data and calculate Rolling Factor Beta.
In this step we're going to filter out stocks that have less than 10 month data, we are doing that because we're going to use Regression for 2 years and stocks that don't have enough data will break our function.

After collecting the valid stocks we calculate rolling factor betas by regressing 1 month return with all the other factors in factor data.