Update Main.py
We have 18 features in our dataset, now we can apply Machine Learning models from here on for each new month we have to form a portfolio with some stocks, and decide which stocks to choose for each portfolio.
This is where we can use machine learning model to predict which stocks to include in the portfolio, if we have a long short portfolio we can predict which stocks to be long and which stocks to be short. We can also use the machine learning model to predict the magnitude of the position in each stock i.e., the weight in the portfolio.

In this case an unsupervised model is used to decide which stocks to use in the portfolio based on grouping, that's why we're going to use the K-Means Clustering Algorithm to keep things simple.

We're going to fit a K-Means Clustering algorithm and split all the stocks in 4 clusters, then we're going to analyze the clusters. K-Means Clustering Algorithm will assign the centroids of the Clusters based on RSI. I've chosen RSI levels of 35, 45, 55, 65 to be the centroids of clusters and then the algorithm will assign a given point to the cluster based on the distance from centroid to that point.

After fitting the model, we can observe that Cluster 3 represents all the stocks with an RSI around 65, Cluster 2 represents RSI around 55, and so on.  Now we can use the Cluster 3 in our portfolio to select a portfolio of stocks every month which have a good momentum and a RSI around 65.