Update Main.py
For each month we'll select assets based on Cluster 3 and form a portfolio based on Efficient Frontier Max Sharpe Ratio

The Hypothesis is that Stocks which have an RSI of around 70 have good momentum and they should carry that momentum even in the next month.

In this step the idea is to create a dictionary with: the first date of the next month, & a list of all the good momentum stocks from the previous month.