Update Main.py
The next step is to define Portfolio Optimization Function. We'll define the function which optimizes portfolio weights using PyPortfolioOpt package and EfficientFrontier optimizer to maximize the Sharpe Ratio.

There may be cases where the optimize_weights function may assign the weight 0 to many stocks or 1 to a single stock to avoid that we'll apply single weight bounds constraint for diversification(minimum half of equal weight and maximum 10% of portfolio).

After that we'll calculate daily returns for each stock which could land up in our potfolio, then loop over each month start, select the stocks for the month and calculate their weights for the next month if the Maximum Sharpe Ratio Optimization fails for a given month, apply equally-weighted weights and calculate each day portfolio return.